Easing specification for react-native docs
`Animated.timing` uses `Easing.inOut` easing by default: https://cl.ly/0L2G0w3v120r.

Without specifying linear easing, you're actually applying easing to the timing—in this example, as it goes from 1 to 5,000. So, if we're using e.g. 60 fps, we'll be showing the correct content at frame 1, frame 150, and frame 300…but everything in between is going to be slightly off.

I didn't check if this also needs to be fixed in the imperative API's implementation details, but you might want to.